Added Gauss function

+++ formula_translation/formula_translation/exponential.c
@@ -1,7 +1,7 @@
 #include "exponential.h"
 
-float exponential(float x)
+double exponential(int x)
 {
-	float res = 1 + x * (1 + x*((1 / 2) + x*((1 / 6) + x*((1 / 24) + (1 / 120)*x))));
+	double res = 1 + x * (1 + x*((1 / 2) + x*((1 / 6) + x*((1 / 24) + (1 / 120)*x))));
 	return res;
 }
